Integrate OCR mechanics into MVVM
