+++ MTTA/Models/TranslationEntry.cs
@@ -16,5 +16,6 @@ namespace MTTA.Models
 
         public string? Original { get; set; }
         public string? Initial { get; set; }
+        public float Confidence { get; set; }
     }
 }
